Rename WinForms project to CpuScheduler
