+++ CpuScheduler/Algorithms.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Windows.Forms;
 
-namespace CpuSchedulingWinForms
+namespace CpuScheduler
 {
     public static class Algorithms
     {
