+++ CpuScheduler/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("CpuSchedulingWinForms")]
+[assembly: AssemblyTitle("CpuScheduler")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("CpuSchedulingWinForms")]
+[assembly: AssemblyProduct("CpuScheduler")]
 [assembly: AssemblyCopyright("Copyright ©  2025 Chris Regan")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
