+++ CpuScheduler/Properties/Settings.settings
